+++ cli/src/tool/tool_register.rs
@@ -79,14 +79,8 @@ pub(crate) async fn register_tool(
         let address = wallet.active_address().map_err(NexusCliError::Any)?;
 
         // Fetch gas and collateral coin objects.
-        let (gas_coin, collateral_coin) = fetch_gas_and_collateral_coins(
-            &sui,
-            conf.sui.net,
-            address,
-            sui_gas_coin,
-            collateral_coin,
-        )
-        .await?;
+        let (gas_coin, collateral_coin) =
+            fetch_gas_and_collateral_coins(&sui, address, sui_gas_coin, collateral_coin).await?;
 
         if gas_coin.coin_object_id == collateral_coin.coin_object_id {
             return Err(NexusCliError::Any(anyhow!(
@@ -244,30 +238,11 @@ pub(crate) async fn register_tool(
 /// if the coins are not present.
 async fn fetch_gas_and_collateral_coins(
     sui: &sui::Client,
-    sui_net: SuiNet,
     addr: sui::Address,
     sui_gas_coin: Option<sui::ObjectID>,
     sui_collateral_coin: Option<sui::ObjectID>,
 ) -> AnyResult<(sui::Coin, sui::Coin), NexusCliError> {
     let mut coins = fetch_all_coins_for_address(sui, addr).await?;
-
-    // We need at least 2 coins. We can create those on Localnet, Devnet and
-    // Testnet.
-    match sui_net {
-        SuiNet::Localnet | SuiNet::Devnet | SuiNet::Testnet if coins.len() < 2 => {
-            // Only call once because on Localnet and Devnet, we get 5 coins and
-            // on Testnet this will be rate-limited.
-            request_tokens_from_faucet(sui_net, addr).await?;
-
-            coins = fetch_all_coins_for_address(sui, addr).await?;
-        }
-        SuiNet::Mainnet if coins.len() < 2 => {
-            return Err(NexusCliError::Any(anyhow!(
-                "The wallet does not have enough coins to register the tool"
-            )));
-        }
-        _ => (),
-    }
 
     if coins.len() < 2 {
         return Err(NexusCliError::Any(anyhow!(